refactor: Extract component from TransactionRow in specific files
